Script to versions and tags (#41)
## WHAT:

1. Read the .version file to get the latest version.
2. Use a selector prompt to choose among "major", "minor", or "patch".
3. Update the version based on the user's choice.
4. Create a new branch with the format new-tag-version-v[version
selected by the user].
5. Update the version in package.json based on the new version.
6. Check if the new version is already released.
7. Update the .version file with the new version.
8. If the version is not released:
9. Commit the new version changes in `package.json` and
`package-lock.json`.
10. Handle the error **"error: gpg failed to sign the data fatal: failed
to write commit object"** and exit the script if this error happens.
11. Create a git tag with the new version.
12. Push the changes and the tag to the remote repository.
13. Create a pull request on GitHub.
14. Echo the pull request link into the terminal.
15. Check if there is an existing draft release, if so, only edit it by
adding the new changes, if not, create a new draft release on GitHub
using the GitHub CLI with automatically generated release notes.
16. Ensure new changes merged into main are included in the draft
release.
17. If there is an exit or error in the script, undo the changes from
`package.json`, `package-lock.json` and `.version`
18. In the create_tag function, check if there is an existing tag for
the new version. If it exists and has a draft release, add the new
changes. If it does not exist, create a new tag.

## OUT OF SCOPE:

Improve to make the script Idempotent, right now we could only used it
once per tag.